+++ Dyalect/Parser/Model/DyCodeModel.cs
@@ -1,12 +1,17 @@
-﻿namespace Dyalect.Parser.Model
+﻿using System.Collections.Generic;
+
+namespace Dyalect.Parser.Model
 {
     public sealed class DyCodeModel
     {
-        public DyCodeModel(DBlock root, string fileName)
+        public DyCodeModel(DBlock root, DImport[] imports, string fileName)
         {
             Root = root;
+            Imports = imports;
             FileName = fileName;
         }
+
+        public DImport[] Imports { get; }
 
         public DBlock Root { get; }